Conflicts:
	posix/gufsm/clfsm/machines/RoboCup2013/SMGetCloseToGoal.machine/Layout.plist
	posix/gufsm/clfsm/machines/RoboCup2013/SMGoalie.machine/State_FindBall_Transition_0.expr
	posix/gufsm/clfsm/machines/RoboCup2013/SMGoalie.machine/State_Start_OnEntry.mm

+++ clfsm/machines/RoboCup2013/SMGoalieSaver.machine/SMGoalieSaver_Variables.h
@@ -9,7 +9,7 @@ unsigned long	historySize;	///<
 unsigned long	headIndex;	///< Index of the 'head' of the ballDifferences array.
 FilteredVisionObject	ball_info;	///< 
 FilteredBallSighting_t	wbSighting;	///< 
-int	interceptX;	///< What side of the robot the ball will be on if it comes in line with it.
+float	interceptX;	///< What side of the robot the ball will be on if it comes in line with it.
 vec2	currentPos;	///< 
 vec2	averageDiff;	///< Average difference vector for ball position.
 MOTION_SDK_Commands_t	wbMotion;	///<
